refactor: improves file structure and naming conventions WIP #19 #29
