+++ src/js/components/create/confirm-where.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { View, Text } from 'react-native';
 import Icon from 'react-native-vector-icons/FontAwesome';
-import Button from '../../common/Button';
-import styles from '../../../style';
+import Button from '../common/Button';
+import styles from '../../style';
 
 const ConfirmEventWhere = ({ eventWhere }) => { // eslint-disable-line react/prop-types
