+++ src/js/components/create/index.js
@@ -1,18 +1,11 @@
 import React from 'react';
 import { View } from 'react-native';
-import EventDetailsHeader from '../general/event-details-header';
 import TopBar from '../event/top-bar';
 import styles from '../../style';
 
-const CreateEvent = ({ title, location, discardEvent, eventDetails, children }) => { // eslint-disable-line reac
-
+const Create = ({ title, discardEvent, eventDetails }) => {
   return (
     <View>
-      <TopBar
-        location={ title }
-        discardEvent={ discardEvent }
-      />
-
       <EventDetailsHeader
         location={ title }
         eventName={ eventDetails.eventName }
@@ -21,7 +14,6 @@ const CreateEvent = ({ title, location, discardEvent, eventDetails, children }) 
       <View style={styles.container} />
     </View>
   );
-
 };
 
-export default CreateEvent;
+export default Create;
