+++ src/js/components/event/top-bar.js
@@ -1,103 +1,94 @@
 import React, { Component } from 'react';
 import { View, Text } from 'react-native';
-import Actions from 'react-native-router-flux';
 import Button from '../common/Button';
 import styles from '../../style';
 
 class TopBar extends Component {
 
-  constructor (props) {
-    super(props);
-  }
-
-  eventType (location) {
-    console.log(location);
-    let pathname = location.split('/').pop();
-    console.log(pathname);
-
-    if (pathname === 'eventdetails') {
-      return 'Create an event';
-    } else {
-      return pathname.charAt(0).toUpperCase() + pathname.slice(1) + '?';
-    }
-  }
-
   cancelEvent () {
     this.props.discardEvent(); //eslint-disable-line
-    Actions.feed();
+    console.log('go to feed');
   }
 
   render () {
     const primaryPath = this.props.location.split('/')[0]; //eslint-disable-line
 
     return (
-      <View style={styles.topBarContainer}>
+      <View style={ styles.topBarContainer }>
         {
           /* feed */
           !this.props.eventID && primaryPath === 'feed' && //eslint-disable-line
-          <View style={styles.rowCentered}>
-            <Text style={styles.title1}>Feed</Text>
+          <View style={ styles.rowCentered }>
+            <Text style={ styles.title1 }>Feed</Text>
           </View>
         }
         {
         /* /create-event */
         !this.props.eventID && primaryPath === 'eventdetails' && //eslint-disable-line
-        <View style={styles.rowSpaced}>
-          <Button
-            textStyle={styles.topBarButtonText}
-            onClick={ () => { Actions.pop(); } }
-          > Back </Button>
-          <Text style={styles.title1}> { this.eventType(this.props.location) }</Text>
-          <Button
-            textStyle={styles.topBarButtonText}
-            onClick={ () => { this.cancelEvent(); } }
-          > Cancel </Button>
-        </View>
+          <View style={ styles.rowSpaced }>
+            <Button
+              textStyle={ styles.topBarButtonText }
+              onClick={ () => { console.log('back'); }}
+            >
+              Back
+            </Button>
+            <Text style={ styles.title1 }>What/Where/When</Text>
+            <Button
+              textStyle={ styles.topBarButtonText }
+              onClick={ () => { this.cancelEvent(); }}
+            > Cancel </Button>
+          </View>
         }
         {
             /* /invite-friends */
-        !this.props.eventID && primaryPath === 'inviteFriends' &&
-        <View style={styles.rowSpaced}>
-          <Button textStyle={styles.topBarButtonText} onClick={ () => { Actions.pop(); } }> Back </Button>
-          <Text style={styles.title1}> Invite Friends </Text>
-          <Button textStyle={styles.topBarButtonText} onClick={ () => { this.cancelEvent(); } }> Cancel </Button>
-        </View>
+          !this.props.eventID && primaryPath === 'inviteFriends' &&
+          <View style={styles.rowSpaced}>
+            <Button textStyle={styles.topBarButtonText} onClick={ () => { console.log('back'); } }>
+              Back
+            </Button>
+            <Text style={styles.title1}>
+              Invite Friends
+            </Text>
+            <Button textStyle={styles.topBarButtonText} onClick={ () => { this.cancelEvent(); } }>
+              Cancel
+            </Button>
+          </View>
         }
         {
             /* /confirm-event */
-        !this.props.eventID && primaryPath === 'confirm' &&
-        <View style={styles.rowSpaced}>
-          <Button
-            textStyle={styles.topBarButtonText}
-            onClick={ () => { Actions.pop(); } }
-          > Back </Button>
-          <Text style={styles.title1}> { primaryPath.charAt(0).toUpperCase() + primaryPath.slice(1) } </Text>
-          <Button textStyle={styles.topBarButtonText} onClick={ () => { this.cancelEvent(); } }> Cancel </Button>
-        </View>
+          !this.props.eventID && primaryPath === 'confirm' &&
+          <View style={styles.rowSpaced}>
+            <Button textStyle={styles.topBarButtonText} onClick={ () => { console.log('back'); }}>
+              Back
+            </Button>
+            <Text style={styles.title1}>
+              { primaryPath.charAt(0).toUpperCase() + primaryPath.slice(1) }
+            </Text>
+            <Button textStyle={styles.topBarButtonText} onClick={ () => { this.cancelEvent(); } }>
+              Cancel
+            </Button>
+          </View>
         }
         {
             /* /calendar */
-        !this.props.eventID && primaryPath === 'albums' &&
+          !this.props.eventID && primaryPath === 'albums' &&
 
-        <View style={styles.rowCentered}>
-          <Button textStyle={styles.topBarButtonText} />
-          <Text style={styles.title1}>
-            { primaryPath.charAt(0).toUpperCase() + primaryPath.slice(1) }
-          </Text>
-          <Button textStyle={styles.topBarButtonText} />
-        </View>
+          <View style={styles.rowCentered}>
+            <Text style={styles.title1}>
+              { primaryPath.charAt(0).toUpperCase() + primaryPath.slice(1) }
+            </Text>
+          </View>
         }
         {
             /* /calendar */
-        !this.props.eventID && primaryPath === 'calendar' &&
+          !this.props.eventID && primaryPath === 'calendar' &&
 
-        <View style={styles.rowCentered}>
-          <Button textStyle={styles.topBarButtonText} />
-          <Text style={styles.title1}>
-            { primaryPath.charAt(0).toUpperCase() + primaryPath.slice(1) }
-          </Text>
-          <Button textStyle={styles.topBarButtonText} />
-        </View>
+          <View style={styles.rowCentered}>
+            <Text style={styles.title1}>
+              { primaryPath.charAt(0).toUpperCase() + primaryPath.slice(1) }
+            </Text>
+            <Button textStyle={ styles.topBarButtonText } />
+          </View>
         }
 
         {
@@ -106,47 +97,41 @@ class TopBar extends Component {
             <Text style={styles.title1}>Profile</Text>
           </View>
         }
-        { /**** EVENT ****/ }
         {
-            //User is Host and its a Poll
-        this.props.eventID && this.props.userIsHost && this.props.isPoll &&
-        <View style={styles.rowSpaced}>
-          <Button textStyle={styles.topBarButtonText}> </Button>
-          <Text style={styles.title1}>Poll</Text>
-          <Button textStyle={styles.topBarButtonText}></Button>
-        </View>
+          this.props.eventID && this.props.userIsHost && this.props.isPoll &&
+          <View style={styles.rowSpaced}>
+            <Text style={styles.title1}>Poll</Text>
+          </View>
         }
         {
-            //User is Invitee and its an Event
-        this.props.eventID && !this.props.userIsHost && !this.props.isPoll &&
-        <View style={styles.rowSpaced}>
-          <Button textStyle={styles.topBarButtonText}> </Button>
-          <Text style={styles.title1}>Event</Text>
-          <Button textStyle={styles.topBarButtonText}></Button>
-        </View>
+          this.props.eventID && !this.props.userIsHost && !this.props.isPoll &&
+          <View style={styles.rowSpaced}>
+            <Text style={styles.title1}>Event</Text>
+          </View>
         }
         {
           // User is Invitee and its a Poll
           this.props.eventID && !this.props.userIsHost && this.props.isPoll &&
-          <View style={styles.rowSpaced}>
-            <Button textStyle={styles.topBarButtonText}> </Button>
-            <Text style={styles.title1}>Poll</Text>
-            <Button textStyle={styles.topBarButtonText}></Button>
+          <View style={ styles.rowSpaced }>
+            <Text style={ styles.title1 }>Poll</Text>
           </View>
         }
         {
           // User is Host and its an Event
-        this.props.eventID && this.props.userIsHost && !this.props.isPoll &&
-        <View style={styles.rowSpaced}>
-          <Button textStyle={styles.topBarButtonText} onClick={ () => { this.props.handleEdit(this.props.event); } } to={ 'edit/' + this.props.eventID }> Edit </Button>
-          <Text style={styles.title1}>Event</Text>
-          <Button
-            textStyle={styles.topBarButtonText}
-            onClick={ this.props.displayCancelModal }
-          >
-            Cancel
-          </Button>
-        </View>
+          this.props.eventID && this.props.userIsHost && !this.props.isPoll &&
+          <View style={styles.rowSpaced}>
+            <Button
+              textStyle={styles.topBarButtonText}
+              onClick={ () => { this.props.handleEdit(this.props.event); } }
+              to={ `edit/ ${this.props.eventID}` }
+            >
+              Edit
+            </Button>
+            <Text style={styles.title1}>Event</Text>
+            <Button textStyle={styles.topBarButtonText} onClick={ this.props.displayCancelModal }>
+              Cancel
+            </Button>
+          </View>
         }
       </View>
     );
