+++ src/js/containers/create/invite.js
@@ -1,0 +1,38 @@
+import { connect } from 'react-redux';
+import { addInvitee, removeInvitee, clearCreateEvent } from '../../actions/create-event';
+import InviteFriends from '../../components/create/invite';
+import jsonState from '../../testState/jsonStateCreate.json';
+
+const mapStateToProps = (state) => { // eslint-disable-line no-unused-vars
+
+  const friends = jsonState.createEvent.friends;
+  const invitees = jsonState.createEvent.invitees;
+
+  return {
+    friends,
+    invitees,
+    eventDetails: jsonState.createEvent.eventDetails
+  };
+};
+
+const mapDispatchToProps = dispatch => ({
+  handleSelected: (id, index) => {
+    dispatch(addInvitee(id, index));
+  },
+
+  removeSelected: (id, index) => {
+    dispatch(removeInvitee(id, index));
+  },
+
+  discardEvent: () => {
+    dispatch(clearCreateEvent());
+  }
+});
+
+
+const InviteFriendsContainer = connect(
+    mapStateToProps,
+    mapDispatchToProps
+)(InviteFriends);
+
+export default InviteFriendsContainer;
